chore: Update WebSocket server code from feedback
